refactor: Create package trajectory
