+++ src/trajectory/trajectory.py
